+++ src/verbose/values/trace_info_value.rs
@@ -24,7 +24,7 @@ impl<'a> TraceInfoValue<'a> {
         buf.try_extend_from_slice(self.value.as_bytes())?;
         if buf.remaining_capacity() > 0 {
             // Safe as capacity is checked earlier
-            unsafe { buf.push_unchecked(u8::from(0)) };
+            unsafe { buf.push_unchecked(0) };
         } else {
             return Err(CapacityError::new(()));
         }
